fix: Ranking failed when mean = 0
Detect this case and provide NaN for the probability density function.  A population with a mean of
0 can't, by definition, have a normal distribution so the pdf calculation fails.  In the rankings
plugin, a mean of 0 probably indicates a category with a single building in rather than a number of
identical buildings. In this case, we simply skip the pdf calculation, return NaN, and let
everything else continue as before.